版本时间:2015.10.10

+++ rd/trunk/Server/ServerCode/hawkjava/GameServer/src/com/hawk/game/entity/MonsterEntity.java
@@ -149,7 +149,7 @@ public class MonsterEntity extends HawkDBEntity {
 		return lazyExp;
 	}
 
-	public void setLazyExp(byte lazyExp) {
+	public void setLazyExp(int lazyExp) {
 		this.lazyExp = lazyExp;
 	}
